remove led mutex
this was causing quite a few issues when trying to use the mutex across 2 threads (which should have worked because its a mutex???).

Crashing only happened when the mutex was locked/unlocked from inside of the LED thread, locking/unlocking from the comms function would then cause the VESC to crash (likely a WDT) if it was used in conjunction with the locking/unlocking inside of the LED thread.

We should definately revisit this but theres zero documentation on mutexes inside of VESC so im at a loss.